refactor: Update discord.js version to 14.16
This commit updates the discord.js version to 14.16. It includes changes to the .gitignore file to ignore lock files, updates to file paths in various modules and listeners, and updates to import statements in several files.

The purpose of this commit is to ensure compatibility with the latest version of discord.js and improve the overall stability and performance of the codebase.